Fix dtype mismatch between backbone and custom scoring head
AutoModel.from_pretrained() without an explicit dtype loaded the
LFM2.5 backbone in the checkpoint's stored bfloat16, while
JevScoringHead's nn.Linear layers default-initialize in float32.
Forward pass then hit RuntimeError: mat1 and mat2 must have the same
dtype (BFloat16 and Float). Explicitly load the backbone in float32
to match the head, which also sidesteps T4's weak bfloat16 support.
Verified locally against the real cached checkpoint: both backbone
and head now report torch.float32.

+++ model/head.py
@@ -32,7 +32,11 @@ class JevScoringHead(nn.Module):
 class JevModel(nn.Module):
     def __init__(self, base_model_name: str = BASE_MODEL_NAME):
         super().__init__()
-        self.backbone = AutoModel.from_pretrained(base_model_name)
+        # dtypeを明示しないと、チェックポイントのbfloat16のまま読み込まれ、
+        # float32で初期化される自作ヘッド(JevScoringHead)との間でdtypeが
+        # 食い違い、F.linearで "mat1 and mat2 must have the same dtype" になる。
+        # T4はbfloat16のネイティブサポートが弱いこともあり、float32に統一する。
+        self.backbone = AutoModel.from_pretrained(base_model_name, dtype=torch.float32)
         hidden_size = self.backbone.config.hidden_size
         self.head = JevScoringHead(hidden_size)
